chemtools-api/pyproject.toml
[tool.poetry]
name = "chemtools-api"
version = "0.1.0"
description = "Chemtools API"
authors = ["silox <edruisl99@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-multipart = "^0.0.10"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
aerich = "^0.7.2"
fastapi-utils = "^0.8.0"
dependency-injector = "^4.43.0"
aiohttp = "^3.11.0"
typing-inspect = "^0.9.0"

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "src/db/migrations"
src_folder = "./src/"
